Use cuserid(3) instead of getlogin(3).
getlogin(3) returns the username of the user owning the controlling terminal of the process. When running pam-test in a CI environment, it's quite possible for the program to run without a controlling terminal at all, which returns NULL, which segfaults pam-test. Instead, cuserid(3) returns the username of the geteuid(2) user.